requirements.txt
flask==2.3.2
python-dotenv==1.0.0
pyjwt==2.7.0
bcrypt==4.0.1